Move getSightingsThisMonth to utils
